Move createExpiryDateFormFieldValues to utility class
