+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CardDetailsElement.kt
@@ -7,6 +7,8 @@ import com.stripe.android.cards.CardAccountRangeRepository
 import com.stripe.android.core.strings.ResolvableString
 import com.stripe.android.model.CardBrand
 import com.stripe.android.ui.core.cbc.CardBrandChoiceEligibility
+import com.stripe.android.ui.core.elements.CardDetailsUtil.getExpiryMonthFormFieldEntry
+import com.stripe.android.ui.core.elements.CardDetailsUtil.getExpiryYearFormFieldEntry
 import com.stripe.android.uicore.elements.IdentifierSpec
 import com.stripe.android.uicore.elements.SectionFieldErrorController
 import com.stripe.android.uicore.elements.SectionMultiFieldElement
@@ -107,40 +109,4 @@ internal class CardDetailsElement(
         }
         return combineAsStateFlow(flows) { it.toList() }
     }
-}
-
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-fun createExpiryDateFormFieldValues(entry: FormFieldEntry): Map<IdentifierSpec, FormFieldEntry> {
-    return mapOf(
-        IdentifierSpec.CardExpMonth to getExpiryMonthFormFieldEntry(entry),
-        IdentifierSpec.CardExpYear to getExpiryYearFormFieldEntry(entry)
-    )
-}
-
-private fun getExpiryMonthFormFieldEntry(entry: FormFieldEntry): FormFieldEntry {
-    var month = -1
-    entry.value?.let { date ->
-        val newString = convertTo4DigitDate(date)
-        if (newString.length == 4) {
-            month = requireNotNull(newString.take(2).toIntOrNull())
-        }
-    }
-
-    return entry.copy(
-        value = month.toString().padStart(length = 2, padChar = '0')
-    )
-}
-
-private fun getExpiryYearFormFieldEntry(entry: FormFieldEntry): FormFieldEntry {
-    var year = -1
-    entry.value?.let { date ->
-        val newString = convertTo4DigitDate(date)
-        if (newString.length == 4) {
-            year = requireNotNull(newString.takeLast(2).toIntOrNull()) + 2000
-        }
-    }
-
-    return entry.copy(
-        value = year.toString()
-    )
 }
